+++ fragmentation_core/src/main/java/me/yokeyword/fragmentation/SupportFragmentDelegate.java
@@ -11,6 +11,7 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
+import android.support.v4.app.FragmentationMagician;
 import android.view.View;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
@@ -22,17 +23,19 @@ import me.yokeyword.fragmentation.helper.internal.TransactionRecord;
 import me.yokeyword.fragmentation.helper.internal.VisibleDelegate;
 
 public class SupportFragmentDelegate {
+    private static final long NOT_FOUND_ANIM_TIME = 300L;
+
     static final int STATUS_UN_ROOT = 0;
     static final int STATUS_ROOT_ANIM_DISABLE = 1;
     static final int STATUS_ROOT_ANIM_ENABLE = 2;
 
     private int mRootStatus = STATUS_UN_ROOT;
 
-    boolean mIsSharedElement;
+    private boolean mIsSharedElement;
     FragmentAnimator mFragmentAnimator;
     AnimatorHelper mAnimHelper;
     boolean mLockAnim;
-    private int mCustomEnterAnim = Integer.MIN_VALUE;
+    private int mCustomEnterAnim = Integer.MIN_VALUE, mCustomExitAnim = Integer.MIN_VALUE, mCustomPopExitAnim = Integer.MIN_VALUE;
 
     private Handler mHandler;
     private boolean mFirstCreateView = true;
@@ -54,6 +57,8 @@ public class SupportFragmentDelegate {
     boolean mAnimByActivity = true;
     EnterAnimListener mEnterAnimListener;
 
+    private boolean mRootViewClickable;
+
     public SupportFragmentDelegate(ISupportFragment support) {
         if (!(support instanceof Fragment))
             throw new RuntimeException("Must extends Fragment");
@@ -69,7 +74,7 @@ public class SupportFragmentDelegate {
         if (mTransactionDelegate == null)
             throw new RuntimeException(mFragment.getClass().getSimpleName() + " not attach!");
 
-        return new ExtraTransaction.ExtraTransactionImpl<>(mSupportF, mTransactionDelegate, false);
+        return new ExtraTransaction.ExtraTransactionImpl<>((FragmentActivity) mSupport, mSupportF, mTransactionDelegate, false);
     }
 
     public void onAttach(Activity activity) {
@@ -91,7 +96,9 @@ public class SupportFragmentDelegate {
             mIsSharedElement = bundle.getBoolean(TransactionDelegate.FRAGMENTATION_ARG_IS_SHARED_ELEMENT, false);
             mContainerId = bundle.getInt(TransactionDelegate.FRAGMENTATION_ARG_CONTAINER);
             mReplaceMode = bundle.getBoolean(TransactionDelegate.FRAGMENTATION_ARG_REPLACE, false);
-            mCustomEnterAnim = bundle.getInt(TransactionDelegate.FRAGMENTATION_ARG_CUSTOM_END_ANIM, Integer.MIN_VALUE);
+            mCustomEnterAnim = bundle.getInt(TransactionDelegate.FRAGMENTATION_ARG_CUSTOM_ENTER_ANIM, Integer.MIN_VALUE);
+            mCustomExitAnim = bundle.getInt(TransactionDelegate.FRAGMENTATION_ARG_CUSTOM_EXIT_ANIM, Integer.MIN_VALUE);
+            mCustomPopExitAnim = bundle.getInt(TransactionDelegate.FRAGMENTATION_ARG_CUSTOM_POP_EXIT_ANIM, Integer.MIN_VALUE);
         }
 
         if (savedInstanceState == null) {
@@ -101,11 +108,43 @@ public class SupportFragmentDelegate {
             mFragmentAnimator = savedInstanceState.getParcelable(TransactionDelegate.FRAGMENTATION_STATE_SAVE_ANIMATOR);
             mIsHidden = savedInstanceState.getBoolean(TransactionDelegate.FRAGMENTATION_STATE_SAVE_IS_HIDDEN);
             mContainerId = savedInstanceState.getInt(TransactionDelegate.FRAGMENTATION_ARG_CONTAINER);
+
+            // RootFragment
+            if (mRootStatus != STATUS_UN_ROOT) {
+                FragmentationMagician.reorderIndices(mFragment.getFragmentManager());
+            }
         }
 
         // Fix the overlapping BUG on pre-24.0.0
         processRestoreInstanceState(savedInstanceState);
         mAnimHelper = new AnimatorHelper(_mActivity.getApplicationContext(), mFragmentAnimator);
+
+        final Animation enter = getEnterAnim();
+        if (enter == null) return;
+
+        getEnterAnim().setAnimationListener(new Animation.AnimationListener() {
+
+            @Override
+            public void onAnimationStart(Animation animation) {
+                mSupport.getSupportDelegate().mFragmentClickable = false;  // 开启防抖动
+
+                mHandler.postDelayed(new Runnable() {
+                    @Override
+                    public void run() {
+                        mSupport.getSupportDelegate().mFragmentClickable = true;
+                    }
+                }, enter.getDuration());
+            }
+
+            @Override
+            public void onAnimationEnd(Animation animation) {
+            }
+
+            @Override
+            public void onAnimationRepeat(Animation animation) {
+
+            }
+        });
     }
 
     public Animation onCreateAnimation(int transit, boolean enter, int nextAnim) {
@@ -155,18 +194,17 @@ public class SupportFragmentDelegate {
 
         View view = mFragment.getView();
         if (view != null) {
+            mRootViewClickable = view.isClickable();
             view.setClickable(true);
             setBackground(view);
         }
+
 
         if (savedInstanceState != null
                 || mRootStatus == STATUS_ROOT_ANIM_DISABLE
                 || (mFragment.getTag() != null && mFragment.getTag().startsWith("android:switcher:"))
                 || (mReplaceMode && !mFirstCreateView)) {
             notifyEnterAnimEnd();
-        } else if (mCustomEnterAnim != Integer.MIN_VALUE) {
-            fixAnimationListener(mCustomEnterAnim == 0 ?
-                    mAnimHelper.getNoneAnim() : AnimationUtils.loadAnimation(_mActivity, mCustomEnterAnim));
         }
 
         if (mFirstCreateView) {
@@ -185,6 +223,7 @@ public class SupportFragmentDelegate {
     public void onDestroyView() {
         mSupport.getSupportDelegate().mFragmentClickable = true;
         getVisibleDelegate().onDestroyView();
+        getHandler().removeCallbacks(mNotifyEnterAnimEndRunnable);
     }
 
     public void onDestroy() {
@@ -200,15 +239,28 @@ public class SupportFragmentDelegate {
     }
 
     /**
-     * If you want to call the start()/pop()/showHideFragment() on the onCreateXX/onActivityCreated,
-     * call this method to deliver the transaction to the queue.
+     * Causes the Runnable r to be added to the action queue.
      * <p>
-     * 在onCreate/onCreateView/onActivityCreated中使用 start()/pop()/showHideFragment(),请使用该方法把你的任务入队
+     * The runnable will be run after all the previous action has been run.
+     * <p>
+     * 前面的事务全部执行后 执行该Action
      *
-     * @param runnable start() , pop() or showHideFragment()
+     * @deprecated Use {@link #post(Runnable)} instead.
      */
+    @Deprecated
     public void enqueueAction(Runnable runnable) {
-        getHandler().postDelayed(runnable, mAnimHelper == null ? 0 : mAnimHelper.enterAnim.getDuration());
+        post(runnable);
+    }
+
+    /**
+     * Causes the Runnable r to be added to the action queue.
+     * <p>
+     * The runnable will be run after all the previous action has been run.
+     * <p>
+     * 前面的事务全部执行后 执行该Action
+     */
+    public void post(final Runnable runnable) {
+        mTransactionDelegate.post(runnable);
     }
 
     /**
@@ -354,7 +406,10 @@ public class SupportFragmentDelegate {
      * 隐藏软键盘
      */
     public void hideSoftInput() {
-        SupportHelper.hideSoftInput(mFragment.getView());
+        Activity activity = mFragment.getActivity();
+        if (activity == null) return;
+        View view = activity.getWindow().getDecorView();
+        SupportHelper.hideSoftInput(view);
     }
 
     /**
@@ -419,10 +474,14 @@ public class SupportFragmentDelegate {
     }
 
     /**
-     * Launch a fragment while poping self.
+     * Start the target Fragment and pop itself
      */
     public void startWithPop(ISupportFragment toFragment) {
-        mTransactionDelegate.dispatchStartTransaction(mFragment.getFragmentManager(), mSupportF, toFragment, 0, ISupportFragment.STANDARD, TransactionDelegate.TYPE_ADD_WITH_POP);
+        mTransactionDelegate.startWithPop(mFragment.getFragmentManager(), mSupportF, toFragment);
+    }
+
+    public void startWithPopTo(ISupportFragment toFragment, Class<?> targetFragmentClass, boolean includeTargetFragment) {
+        mTransactionDelegate.startWithPopTo(mFragment.getFragmentManager(), mSupportF, toFragment, targetFragmentClass.getName(), includeTargetFragment);
     }
 
     public void replaceFragment(ISupportFragment toFragment, boolean addToBackStack) {
@@ -442,7 +501,7 @@ public class SupportFragmentDelegate {
     }
 
     public void startChildWithPop(ISupportFragment toFragment) {
-        mTransactionDelegate.dispatchStartTransaction(getChildFragmentManager(), getTopFragment(), toFragment, 0, ISupportFragment.STANDARD, TransactionDelegate.TYPE_ADD_WITH_POP);
+        mTransactionDelegate.startWithPop(getChildFragmentManager(), getTopFragment(), toFragment);
     }
 
     public void replaceChildFragment(ISupportFragment toFragment, boolean addToBackStack) {
@@ -450,14 +509,14 @@ public class SupportFragmentDelegate {
     }
 
     public void pop() {
-        mTransactionDelegate.back(mFragment.getFragmentManager());
+        mTransactionDelegate.pop(mFragment.getFragmentManager());
     }
 
     /**
      * Pop the child fragment.
      */
     public void popChild() {
-        mTransactionDelegate.back(getChildFragmentManager());
+        mTransactionDelegate.pop(getChildFragmentManager());
     }
 
     /**
@@ -470,7 +529,6 @@ public class SupportFragmentDelegate {
      * @param includeTargetFragment 是否包含该fragment
      */
     public void popTo(Class<?> targetFragmentClass, boolean includeTargetFragment) {
-        getChildFragmentManager().popBackStack();
         popTo(targetFragmentClass, includeTargetFragment, null);
     }
 
@@ -498,6 +556,10 @@ public class SupportFragmentDelegate {
         mTransactionDelegate.popTo(targetFragmentClass.getName(), includeTargetFragment, afterPopTransactionRunnable, getChildFragmentManager(), popAnim);
     }
 
+    public void popQuiet() {
+        mTransactionDelegate.popQuiet(mFragment.getFragmentManager());
+    }
+
     private FragmentManager getChildFragmentManager() {
         return mFragment.getChildFragmentManager();
     }
@@ -519,14 +581,9 @@ public class SupportFragmentDelegate {
     }
 
     private void fixAnimationListener(Animation enterAnim) {
-        mSupport.getSupportDelegate().mFragmentClickable = false;
         // AnimationListener is not reliable.
-        getHandler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                notifyEnterAnimEnd();
-            }
-        }, enterAnim.getDuration());
+        getHandler().postDelayed(mNotifyEnterAnimEndRunnable, enterAnim.getDuration());
+        mSupport.getSupportDelegate().mFragmentClickable = true;
 
         if (mEnterAnimListener != null) {
             getHandler().post(new Runnable() {
@@ -538,6 +595,30 @@ public class SupportFragmentDelegate {
             });
         }
     }
+
+    private Runnable mNotifyEnterAnimEndRunnable = new Runnable() {
+        @Override
+        public void run() {
+            if (mFragment == null) return;
+            mSupportF.onEnterAnimationEnd(mSaveInstanceState);
+
+            if (mRootViewClickable) return;
+            final View view = mFragment.getView();
+            if (view == null) return;
+            ISupportFragment preFragment = SupportHelper.getPreFragment(mFragment);
+            if (preFragment == null) return;
+
+            long prePopExitDuration = preFragment.getSupportDelegate().getPopExitAnimDuration();
+            long enterDuration = getEnterAnimDuration();
+
+            mHandler.postDelayed(new Runnable() {
+                @Override
+                public void run() {
+                    view.setClickable(false);
+                }
+            }, prePopExitDuration - enterDuration);
+        }
+    };
 
     private void compatSharedElements() {
         notifyEnterAnimEnd();
@@ -569,13 +650,7 @@ public class SupportFragmentDelegate {
     }
 
     private void notifyEnterAnimEnd() {
-        getHandler().post(new Runnable() {
-            @Override
-            public void run() {
-                if (mFragment == null) return;
-                mSupportF.onEnterAnimationEnd(mSaveInstanceState);
-            }
-        });
+        getHandler().post(mNotifyEnterAnimEndRunnable);
         mSupport.getSupportDelegate().mFragmentClickable = true;
     }
 
@@ -595,6 +670,77 @@ public class SupportFragmentDelegate {
 
     public FragmentActivity getActivity() {
         return _mActivity;
+    }
+
+    private Animation getEnterAnim() {
+        if (mCustomEnterAnim == Integer.MIN_VALUE) {
+            if (mAnimHelper != null && mAnimHelper.enterAnim != null) {
+                return mAnimHelper.enterAnim;
+            }
+        } else {
+            try {
+                return AnimationUtils.loadAnimation(_mActivity, mCustomEnterAnim);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return null;
+    }
+
+    private long getEnterAnimDuration() {
+        Animation enter = getEnterAnim();
+        if (enter != null) {
+            return enter.getDuration();
+        }
+        return NOT_FOUND_ANIM_TIME;
+    }
+
+    public long getExitAnimDuration() {
+        if (mCustomExitAnim == Integer.MIN_VALUE) {
+            if (mAnimHelper != null && mAnimHelper.exitAnim != null) {
+                return mAnimHelper.exitAnim.getDuration();
+            }
+        } else {
+            try {
+                return AnimationUtils.loadAnimation(_mActivity, mCustomExitAnim).getDuration();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+
+        }
+        return NOT_FOUND_ANIM_TIME;
+    }
+
+    private long getPopExitAnimDuration() {
+        if (mCustomPopExitAnim == Integer.MIN_VALUE) {
+            if (mAnimHelper != null && mAnimHelper.popExitAnim != null) {
+                return mAnimHelper.popExitAnim.getDuration();
+            }
+        } else {
+            try {
+                return AnimationUtils.loadAnimation(_mActivity, mCustomPopExitAnim).getDuration();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+
+        }
+        return NOT_FOUND_ANIM_TIME;
+    }
+
+    @Nullable
+    Animation getExitAnim() {
+        if (mCustomExitAnim == Integer.MIN_VALUE) {
+            if (mAnimHelper != null && mAnimHelper.exitAnim != null) {
+                return mAnimHelper.exitAnim;
+            }
+        } else {
+            try {
+                return AnimationUtils.loadAnimation(_mActivity, mCustomExitAnim);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return null;
     }
 
     interface EnterAnimListener {