+++ fragmentation_core/src/main/java/me/yokeyword/fragmentation/helper/internal/VisibleDelegate.java
@@ -6,7 +6,7 @@ import android.os.Looper;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentationHack;
+import android.support.v4.app.FragmentationMagician;
 
 import java.util.List;
 
@@ -25,7 +25,6 @@ public class VisibleDelegate {
     private boolean mNeedDispatch = true;
     private boolean mInvisibleWhenLeave;
     private boolean mIsFirstVisible = true;
-    private boolean mFixStatePagerAdapter;
     private boolean mFirstCreateViewCompatReplace = true;
 
     private Handler mHandler;
@@ -42,10 +41,9 @@ public class VisibleDelegate {
     public void onCreate(@Nullable Bundle savedInstanceState) {
         if (savedInstanceState != null) {
             mSaveInstanceState = savedInstanceState;
-            if (!mFixStatePagerAdapter) { // setUserVisibleHint() may be called before onCreate()
-                mInvisibleWhenLeave = savedInstanceState.getBoolean(FRAGMENTATION_STATE_SAVE_IS_INVISIBLE_WHEN_LEAVE);
-                mFirstCreateViewCompatReplace = savedInstanceState.getBoolean(FRAGMENTATION_STATE_SAVE_COMPAT_REPLACE);
-            }
+            // setUserVisibleHint() may be called before onCreate()
+            mInvisibleWhenLeave = savedInstanceState.getBoolean(FRAGMENTATION_STATE_SAVE_IS_INVISIBLE_WHEN_LEAVE);
+            mFirstCreateViewCompatReplace = savedInstanceState.getBoolean(FRAGMENTATION_STATE_SAVE_COMPAT_REPLACE);
         }
     }
 
@@ -63,8 +61,7 @@ public class VisibleDelegate {
             mFirstCreateViewCompatReplace = false;
         }
 
-        if (!mInvisibleWhenLeave && !mFragment.isHidden() &&
-                (mFragment.getUserVisibleHint() || mFixStatePagerAdapter)) {
+        if (!mInvisibleWhenLeave && !mFragment.isHidden() && mFragment.getUserVisibleHint()) {
             if ((mFragment.getParentFragment() != null && isFragmentVisible(mFragment.getParentFragment()))
                     || mFragment.getParentFragment() == null) {
                 mNeedDispatch = false;
@@ -107,19 +104,15 @@ public class VisibleDelegate {
 
     public void onDestroyView() {
         mIsFirstVisible = true;
-        mFixStatePagerAdapter = false;
     }
 
     public void setUserVisibleHint(boolean isVisibleToUser) {
-        if (mFragment.isResumed() || (mFragment.isDetached() && isVisibleToUser)) {
+        if (mFragment.isResumed() || (!mFragment.isAdded() && isVisibleToUser)) {
             if (!mIsSupportVisible && isVisibleToUser) {
                 safeDispatchUserVisibleHint(true);
             } else if (mIsSupportVisible && !isVisibleToUser) {
                 dispatchSupportVisible(false);
             }
-        } else if (isVisibleToUser) {
-            mInvisibleWhenLeave = false;
-            mFixStatePagerAdapter = true;
         }
     }
 
@@ -142,6 +135,8 @@ public class VisibleDelegate {
     }
 
     private void dispatchSupportVisible(boolean visible) {
+        if (visible && isParentInvisible()) return;
+
         if (mIsSupportVisible == visible) {
             mNeedDispatch = true;
             return;
@@ -149,33 +144,49 @@ public class VisibleDelegate {
 
         mIsSupportVisible = visible;
 
-        if (!mNeedDispatch) {
-            mNeedDispatch = true;
-        } else {
-            if (!mFragment.isAdded()) return;
-            FragmentManager fragmentManager = mFragment.getChildFragmentManager();
-            if (fragmentManager != null) {
-                List<Fragment> childFragments = FragmentationHack.getActiveFragments(fragmentManager);
-                if (childFragments != null) {
-                    for (Fragment child : childFragments) {
-                        if (child instanceof ISupportFragment && !child.isHidden() && child.getUserVisibleHint()) {
-                            ((ISupportFragment) child).getSupportDelegate().getVisibleDelegate().dispatchSupportVisible(visible);
-                        }
-                    }
-                }
-            }
-        }
-
         if (visible) {
+            if (checkAddState()) return;
             mSupportF.onSupportVisible();
 
             if (mIsFirstVisible) {
                 mIsFirstVisible = false;
                 mSupportF.onLazyInitView(mSaveInstanceState);
             }
+            dispatchChild(true);
         } else {
+            dispatchChild(false);
             mSupportF.onSupportInvisible();
         }
+    }
+
+    private void dispatchChild(boolean visible) {
+        if (!mNeedDispatch) {
+            mNeedDispatch = true;
+        } else {
+            if (checkAddState()) return;
+            FragmentManager fragmentManager = mFragment.getChildFragmentManager();
+            List<Fragment> childFragments = FragmentationMagician.getActiveFragments(fragmentManager);
+            if (childFragments != null) {
+                for (Fragment child : childFragments) {
+                    if (child instanceof ISupportFragment && !child.isHidden() && child.getUserVisibleHint()) {
+                        ((ISupportFragment) child).getSupportDelegate().getVisibleDelegate().dispatchSupportVisible(visible);
+                    }
+                }
+            }
+        }
+    }
+
+    private boolean isParentInvisible() {
+        ISupportFragment fragment = (ISupportFragment) mFragment.getParentFragment();
+        return fragment != null && !fragment.isSupportVisible();
+    }
+
+    private boolean checkAddState() {
+        if (!mFragment.isAdded()) {
+            mIsSupportVisible = !mIsSupportVisible;
+            return true;
+        }
+        return false;
     }
 
     private boolean isFragmentVisible(Fragment fragment) {